requirements.txt
python-dotenv==1.0.0
streamlit==1.29.0
requests==2.31.0
pandas==2.1.4
streamlit-folium==0.18.0
pydantic==2.5.3
langchain==0.0.353
tiktoken==0.5.2
openai==1.6.1
faiss-cpu==1.7.4
transformers==4.36.2
sentence_transformers==2.2.2
rapidfuzz==3.6.1
